Changed away from the class style back to modules
Modules still do not adhere to interfaces so using a simple trick at the
end of the module to enforce the interface checking against the module
declaration. Not the best choice but works.

Refactored the modules back in instead of classes. Kind of would have
wanted the classes but oh well... Such is life.